backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.2
Pillow==10.1.0
pytesseract==0.3.10
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.1
statsmodels==0.14.1
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
python-dateutil==2.8.2
Werkzeug==3.0.1
cryptography
openpyxl